Add support for large graphs

+++ src/ch_constructor.h
@@ -562,7 +562,7 @@ std::vector<int> CHConstructor<NodeT, EdgeT>::calcEdgeDiffs(std::vector<NodeID> 
 	uint size(nodes.size());
 	#pragma omp parallel for num_threads(_num_threads) schedule(dynamic)
 	for (uint i = 0; i<size; i++) {
-		edge_diffs[i] = shortcuts[i].size() - (int) _base_graph.getNrOfEdges(nodes[i]);
+		edge_diffs[i] = (int)shortcuts[i].size() - (int)_base_graph.getNrOfEdges(nodes[i]);
 	}
 
 	return edge_diffs;
